Adding images

+++ database/migrations/2025_02_17_065330_create_orders_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->foreignIdFor(User::class)->constrained()->cascadeOnDelete();
-            $table->unsignedInteger('total_price');
+            $table->decimal('total_price', 10, 2);
             $table->enum("status",["pending","completed","canceled"])->default("pending");
             $table->timestamps();
         });
